refactor: 상품 등록 페이지  타입 폴더 구조 변경
