--- src/utils/types/auction/auctionEnroll.ts
@@ -1,14 +1,0 @@
-export interface AuctionEnrollRequest {
-  title: string;
-  productCategory: string;
-  initPrice: number;
-  endDate: string;
-  productStatus: string;
-  purchaseTime: string;
-  description: string;
-  tradeMethod: string;
-  imageUrls: string[];
-  si: string;
-  gu: string;
-  dong: string;
-}
