--- src/utils/types/image/imageController.ts
@@ -1,3 +1,0 @@
-export interface ImageControllerRequest {
-  imageUrls: string[];
-}
